+++ src/main-v1.pl
@@ -138,14 +138,6 @@ sub parseQuery {
             $next_command = ["addlinenum", undef, undef];
             $last_command = $a;
 
-        } elsif ($a eq "addlinenum2") {
-            $next_command = ["addlinenum2", undef];
-            $last_command = $a;
-
-        } elsif ($a eq "addnumsortable") {
-            $next_command = ["addnumsortable", undef, undef];
-            $last_command = $a;
-
         } elsif ($a eq "addcross") {
             $next_command = ["addcross", undef, undef];
             $last_command = $a;
@@ -454,46 +446,6 @@ sub parseQuery {
                 } elsif (!defined($curr_command->[2])) {
                     my $value = $a;
                     $curr_command->[2] = $value;
-                } else {
-                    die "Unknown argument: $a";
-                }
-
-            } elsif ($curr_command->[0] eq "addlinenum2") {
-                if ($a eq "--name") {
-                    die "option $a needs an argument" unless (@$argv);
-                    my $name = shift(@$argv);
-                    if (defined($curr_command->[1])) {
-                        die "duplicated option: --name";
-                    }
-                    $curr_command->[1] = $name;
-                } elsif (!defined($curr_command->[1])) {
-                    my $name = $a;
-                    $curr_command->[1] = $name;
-                } else {
-                    die "Unknown argument: $a";
-                }
-
-            } elsif ($curr_command->[0] eq "addnumsortable") {
-                if ($a eq "--name") {
-                    die "option $a needs an argument" unless (@$argv);
-                    my $name = shift(@$argv);
-                    if (defined($curr_command->[1])) {
-                        die "duplicated option: --name";
-                    }
-                    $curr_command->[1] = $name;
-                } elsif ($a eq "--col") {
-                    die "option $a needs an argument" unless (@$argv);
-                    my $col = shift(@$argv);
-                    if (defined($curr_command->[2])) {
-                        die "duplicated option: --col";
-                    }
-                    $curr_command->[2] = $col;
-                } elsif (!defined($curr_command->[1])) {
-                    my $name = $a;
-                    $curr_command->[1] = $name;
-                } elsif (!defined($curr_command->[2])) {
-                    my $col = $a;
-                    $curr_command->[2] = $col;
                 } else {
                     die "Unknown argument: $a";
                 }
@@ -869,19 +821,6 @@ sub parseQuery {
                 $c->[2] = 1;
             }
             push(@$commands2, ["addlinenum", $c->[1], $c->[2]]);
-        } elsif ($c->[0] eq "addlinenum2") {
-            if (!defined($c->[1])) {
-                die "subcommand \`addlinenum2\` needs --name option";
-            }
-            push(@$commands2, ["addlinenum2", $c->[1]]);
-        } elsif ($c->[0] eq "addnumsortable") {
-            if (!defined($c->[1])) {
-                die "subcommand \`addnumsortable\` needs --name option";
-            }
-            if (!defined($c->[2])) {
-                die "subcommand \`addnumsortable\` needs --col option";
-            }
-            push(@$commands2, ["addnumsortable", $c->[1], $c->[2]]);
         } elsif ($c->[0] eq "removecol") {
             if (!defined($c->[1])) {
                 die "subcommand \`removecol\` needs --count option";
@@ -989,8 +928,11 @@ sub parseSortParams {
     my ($args) = @_;
     my @args = @$args;
     my $commands = [];
-    push(@$commands, ["addlinenum2", ""]);
-    my $c = 1;
+    my $c = 0;
+    if (@args) {
+        push(@$commands, ["addlinenum2"]);
+        $c++;
+    }
     while (@args) {
         my $a = pop(@args);
         if ($a =~ /\A([_0-9a-zA-Z][-_0-9a-zA-Z]*):n\z/) {
@@ -1001,7 +943,9 @@ sub parseSortParams {
         $c++;
     }
     push(@$commands, ["sort"]);
-    push(@$commands, ["removecol", $c]);
+    if ($c > 0) {
+        push(@$commands, ["removecol", $c]);
+    }
     $commands;
 }
 
@@ -1448,8 +1392,7 @@ sub build_ircode_command {
             push(@$ircode, ["cmd", "perl \$TOOL_DIR/addlinenum.pl --name $name --value $value"]);
 
         } elsif ($command eq "addlinenum2") {
-            my $name  = escape_for_bash($t->[1]);
-            push(@$ircode, ["cmd", "perl \$TOOL_DIR/addlinenum2.pl --name $name"]);
+            push(@$ircode, ["cmd", "perl \$TOOL_DIR/addlinenum2.pl --name ''"]);
 
         } elsif ($command eq "addnumsortable") {
             my $name  = escape_for_bash($t->[1]);