+++ test/case-bom.sh
@@ -1,3 +1,4 @@
 
 ./xsvutils ./test/data/sample-bom.csv > ./var/test/case-bom-1.tsv
+./xsvutils -v2 ./test/data/prefecture.tsv.gz > ./var/test/gz-1.tsv