+++ test/case-newline.sh
@@ -1,0 +1,8 @@
+
+./xsvutils -v2 ./test/data/newline-unix.tsv > ./var/test/newline-unix.tsv.txt
+./xsvutils -v2 ./test/data/newline-unix.csv > ./var/test/newline-unix.csv.txt
+./xsvutils -v2 ./test/data/newline-dos.tsv > ./var/test/newline-dos.tsv.txt
+./xsvutils -v2 ./test/data/newline-dos.csv > ./var/test/newline-dos.csv.txt
+./xsvutils -v2 ./test/data/newline-mac.tsv > ./var/test/newline-mac.tsv.txt
+./xsvutils -v2 ./test/data/newline-mac.csv > ./var/test/newline-mac.csv.txt
+