+++ test/case-sort.sh
@@ -1,0 +1,6 @@
+
+./xsvutils -v2 ./test/data/sort.tsv cut name,num sort > ./var/test/sort-1.tsv
+./xsvutils -v2 ./test/data/sort.tsv sort name         > ./var/test/sort-2.tsv
+./xsvutils -v2 ./test/data/sort.tsv sort num          > ./var/test/sort-3.tsv
+./xsvutils -v2 ./test/data/sort.tsv sort num:n        > ./var/test/sort-4.tsv
+